hotfix: オーバーフローが起こる場合にモジュロ演算が誤作動するバグを修正

+++ src/matsu/num/transform/fft/number/PrimitiveRootOfPrimeNumber.java
@@ -5,7 +5,7 @@
  * http://opensource.org/licenses/mit-license.php
  */
 /*
- * 2024.12.26
+ * 2025.8.19
  */
 package matsu.num.transform.fft.number;
 
@@ -52,10 +52,6 @@ public final class PrimitiveRootOfPrimeNumber {
 
     private int[] calcExpMapping() {
 
-        /*
-         * TODO: aの積剰余を計算することにおける, オーバーフローリスクを調査する.
-         */
-
         int[] mapping = new int[this.prime - 1];
 
         final int a = this.primitiveRoot;
@@ -64,7 +60,7 @@ public final class PrimitiveRootOfPrimeNumber {
         int a_n = 1;
         for (int i = 0; i < mapping.length; i++) {
             mapping[i] = a_n;
-            a_n = a_n * a % p;
+            a_n = (int) ((long) a_n * a % p);
         }
 
         return mapping;
